Added the data classses for getting the 16-day forecasts.

+++ app/src/main/java/com/lawal/banji/yahewa/model/DailyForecast.kt
@@ -3,8 +3,8 @@ package com.lawal.banji.yahewa.model
 import com.google.gson.annotations.SerializedName
 
 data class DailyForecast (
-    @SerializedName("sunrise") val sunrise: Double,
-    @SerializedName("sunset") val sunset: Double,
+    @SerializedName("sunrise") val sunrise: Long,
+    @SerializedName("sunset") val sunset: Long,
     @SerializedName("pressure") val pressure: Double,
     @SerializedName("humidity") val humidity: Double,
     @SerializedName("speed") val windSpeed: Double,
